+++ db.json
@@ -25,19 +25,9 @@
   ],
   "transactions": [
     {
-      "transId": "3chwq3FIN",
+      "transId": "SRsi7msiy",
       "userId": "PEAtqHQ3K",
-      "bookId": "PM2GbD63q",
-      "name": "Võ Ngọc Giàu",
-      "info": "Student",
-      "id": "PM2GbD63q",
-      "title": "book 1",
-      "des": "hello"
-    },
-    {
-      "transId": "WqAGcwZfP",
-      "userId": "ThdF7vtxD",
-      "bookId": "O9EBRmOs-"
+      "bookId": "PM2GbD63q"
     }
   ]
 }
