+++ views/books/update.pug
@@ -1,7 +1,7 @@
 h1 Update Title
 form(action='/books/update', method='POST')
 	label Title: 
-	input(name="title", type='text')
+	input(name="title", type='text', value= book.title)
 	br
 	br
 	button Update
