+++ views/transactions/index.pug
@@ -8,7 +8,7 @@ br
 br
 table(style='width:100%', border='1')
     tr
-      th Transaction
+      th Transaction's ID
       th User's name
       th Book
     each transaction in transactions
